itemsコントローラー修正

+++ app/controllers/items_controller.rb
@@ -35,10 +35,12 @@ class ItemsController < ApplicationController
   end
 
   def destroy
-    if @item.destroy
-      redirect_to root_path
-    else
-      render :show
+    if user_signed_in? && (current_user.id == @item.user.id)
+      if @item.destroy
+        redirect_to root_path
+      else
+        render :show
+      end
     end
   end
